endpoint cleanup

+++ internal/api/delete_project_secret_v1alpha1.go
@@ -6,15 +6,16 @@ import (
 
 	"connectrpc.com/connect"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 
+	kargoapi "github.com/akuity/kargo/api/v1alpha1"
 	svcv1alpha1 "github.com/akuity/kargo/pkg/api/service/v1alpha1"
 )
 
-func (s *server) DeleteSecrets(
+func (s *server) DeleteProjectSecret(
 	ctx context.Context,
-	req *connect.Request[svcv1alpha1.DeleteSecretsRequest],
-) (*connect.Response[svcv1alpha1.DeleteSecretsResponse], error) {
+	req *connect.Request[svcv1alpha1.DeleteProjectSecretRequest],
+) (*connect.Response[svcv1alpha1.DeleteProjectSecretResponse], error) {
 	// Check if secret management is enabled
 	if !s.cfg.SecretManagementEnabled {
 		return nil, connect.NewError(connect.CodeUnimplemented, errSecretManagementDisabled)
@@ -34,15 +35,33 @@ func (s *server) DeleteSecrets(
 		return nil, err
 	}
 
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
+	secret := &corev1.Secret{}
+	if err := s.client.Get(
+		ctx,
+		types.NamespacedName{
 			Namespace: project,
 			Name:      name,
 		},
+		secret,
+	); err != nil {
+		return nil, fmt.Errorf("get secret: %w", err)
 	}
+	if secret.Labels[kargoapi.ProjectSecretLabelKey] != kargoapi.LabelTrueValue {
+		return nil, connect.NewError(
+			connect.CodeNotFound,
+			fmt.Errorf(
+				"secret %s/%s exists, but is not labeled with %s=%s",
+				secret.Namespace,
+				secret.Name,
+				kargoapi.ProjectSecretLabelKey,
+				kargoapi.LabelTrueValue,
+			),
+		)
+	}
+
 	if err := s.client.Delete(ctx, secret); err != nil {
 		return nil, fmt.Errorf("delete secret: %w", err)
 	}
 
-	return connect.NewResponse(&svcv1alpha1.DeleteSecretsResponse{}), nil
+	return connect.NewResponse(&svcv1alpha1.DeleteProjectSecretResponse{}), nil
 }
